[tests.vda] test_recode_criteria: Removed the irrelevant-criterion part.
It looks like there isn't a general way to get this invariance for Artiruno without adding a lot of items to the preorder, which is impractical for big item spaces. I could add special handling of the case where there are criteria on which all alternatives are the same, but this wouldn't be useful in practice, because generally, users won't include such criteria.